+++ src/components/Pages/Visualization.js
@@ -4,12 +4,9 @@ import {GiCorn} from "react-icons/gi";
 import {AiFillAndroid} from "react-icons/ai";
 import {SiAdblock} from "react-icons/si";
 import {GrHorton} from "react-icons/gr";
-import {ConcernsMenu} from '../Visualizations/Concerns_Vis/ConcernsMenu'
-import { useData, getFarmersCrops } from '../Visualizations/UseData';
-import {CropPercentages} from '../Visualizations/CropPercentages';
-import {AcresManagedMenu} from '../Visualizations/AcresManagedVis/AcresManagedMenu';
-import { PriorityMenu } from '../Visualizations/PrioritySatisfaction/PriorityMenu';
-import {AffectMenu} from '../Visualizations/AffectProduction/AffectMenu';
+import { useData } from '../Visualizations/UseData';
+import {VisMenu} from '../Visualizations/Menu';
+import { CropPercentages } from '../Visualizations/Questions/CropPercentages';
 
 export const Visualizations = () => {
     return(
@@ -27,19 +24,19 @@ export const Visualizations = () => {
             </Menu>
             </ProSidebar>
         </div> 
-        {<ConcernsMenu dataset={useData('./data/Grower_Crop_Data.csv')}/>}
+        {<VisMenu dataset={useData('./data/Grower_Crop_Data.csv')} vis={"Concerns"}/>}
         <h2>Of these acres, in a given year estimate approximately what percentage are in the following categories?</h2>
 
         {<CropPercentages dataset={useData('./data/Grower_Crop_Data.csv')}/>}
              
         {/*Question written in AcresManagedBarChart*/}
-        {<AcresManagedMenu dataset={useData('./data/Filtered_Crop_Data.csv')}/>}
+        {<VisMenu dataset={useData('./data/Filtered_Crop_Data.csv')} vis={"AcresManaged"}/>}
 
         <h2>Rate your current level of importance/satisfaction with UCCE's delivery of information on these topics, (1-3)</h2>
-        {<PriorityMenu dataset={useData('./data/Grower_Crop_Data.csv')}/>}
+        {<VisMenu dataset={useData('./data/Grower_Crop_Data.csv')} vis={"PrioritySatisfaction"}/>}
 
         <h2>How often do the following priorities affect your recommendations for field crop production?</h2>
-        {<AffectMenu dataset={useData('./data/Grower_Crop_Data.csv')}/>}
+        {<VisMenu dataset={useData('./data/Grower_Crop_Data.csv')} vis={"Affect"}/>}
 
       </div>
     );
