+++ src/components/Visualizations/Questions/AcresManaged.js
